requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
marshmallow==3.22.0
flask-marshmallow==1.2.1
flask-sqlalchemy==3.0.4
sendgrid==6.11.0
marshmallow-sqlalchemy==1.1.0
pyjwt==2.6.0
python-dotenv==1.0.0
Flask-Cors==5.0.0
celery==5.4.0
redis==5.0.8
pytz==2024.2
